+++ fidelius/gateway/paramstore/_paramstoreadmin.py
@@ -1,14 +1,16 @@
 __all__ = [
     'ParameterStoreAdmin',
 ]
-from .paramstore import *
+
 from fidelius.structs import *
+from fidelius.gateway.interface import *
+from ._paramstorerepo import *
 
 import logging
 log = logging.getLogger(__name__)
 
 
-class ParameterStoreAdmin(ParameterStore):
+class ParameterStoreAdmin(IFideliusAdminRepo, ParameterStore):
     def __init__(self, app: str, group: str, env: str, owner: str, finance: str = 'COST', **extra_tags):
         super().__init__(app, group, env)
         self._tags = Tags(application=self._app, owner=owner, tier=env, finance=finance, **extra_tags)