+++ tests/test_some_basic_stuff.py
@@ -1,0 +1,7 @@
+import unittest
+
+from fidelius.gateway.mock import *
+
+
+class TestSomeBasicStuff(unittest.TestCase):
+    pass