+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - ?
+
+### Added
+
+- An interface for Fidelius gateway repos and admins to fulfil
+- A mock implementation of a Fidelius gateway repo and admin that can read 
+  in mock values from JSON files and mess with them in-memory, for use in 
+  unit-tests of other code that uses Fidelius
+
 
 ## [0.6.0] - 2024-04-05